requirements.txt
# FastAPI и зависимости
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Pydantic (новая версия для Python 3.13)
pydantic>=2.9.2
pydantic-settings>=2.5.0

# HTTP клиенты
httpx>=0.27.0
requests>=2.32.0

# AI
google-generativeai>=0.8.0

# Утилиты
python-dotenv>=1.0.1
python-multipart>=0.0.12

